web/fillgraphdb/requirements.txt
pymysql
azure-mgmt-resourcegraph
azure-mgmt-resource
azure-cli-core
azure.identity
requests